Don't delete wheels from external repos with pip_parse (#458)

+++ python/pip_install/extract_wheels/lib/bazel.py
@@ -248,6 +248,7 @@ def extract_wheel(
         )
         build_file.write(contents)
 
-    os.remove(whl.path)
+    if not incremental:
+        os.remove(whl.path)
 
     return "//%s" % directory
